+++ src/modules/task.js
@@ -1,7 +1,7 @@
 // -- Class constructor of the task object
 
 export default class Task {
-  constructor(title, description, dueDate, index,important) {
+  constructor(title, description, dueDate, index, important) {
     this.title = title;
     this.description = description;
     this.dueDate = dueDate;
